Restore libGenetic compilability

+++ src/Spectre.libGenetic/Individual.cpp
@@ -12,7 +12,7 @@ Individual::Individual(long size)
 	this->size = size;
 	spectres = new vector<bool>(size, 0);
 	int i, randomval;
-	srand(time(0));
+	srand(static_cast<unsigned int>(time(0)));
 	for (i = 0; i<size; i++)
 	{
 		randomval = rand() % 2;
@@ -25,7 +25,7 @@ Individual::Individual(long size, long amount)
 	this->size = size;
 	spectres = new vector<bool>(size, 0);
 	int i, randomval;
-	srand(time(0));
+    srand(static_cast<unsigned int>(time(0)));
 	for (i = 0; i<size; i++)
 	{
 		spectres->push_back(false);
@@ -42,7 +42,7 @@ Individual::Individual(long size, long amountFrom, long amountTo)
 	this->size = size;
 	spectres = new vector<bool>(size, 0);
 	int i, randomval, randomTrueVal;
-	srand(time(0));
+    srand(static_cast<unsigned int>(time(0)));
 	randomTrueVal = (rand() % (amountTo - amountFrom + 1)) + amountFrom;
 	for (i = 0; i<size; i++)
 	{
